test: separate static domain ownership validation
