+++ frontend/e2e/domain-links.spec.ts
@@ -1,10 +1,4 @@
 import { test, expect } from '@playwright/test';
-import * as fs from 'fs';
-import * as path from 'path';
-import { fileURLToPath } from 'url';
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const repoRoot = path.resolve(__dirname, '../..');
 
 test.describe('Banana Slides official domains', () => {
   test.beforeEach(async ({ page }) => {
@@ -26,32 +20,4 @@ test.describe('Banana Slides official domains', () => {
     await expect(page.locator(`a[href*="${thirdPartyHost}"]`)).toHaveCount(0);
   });
 
-  test('repository links keep first-party and AI provider domains separate', () => {
-    const firstPartyFiles = [
-      '.github/ISSUE_TEMPLATE/bug_report.yml',
-      'README.md',
-      'README_EN.md',
-      'docs/quickstart.mdx',
-      'docs/zh/quickstart.mdx',
-      'skills/banana-cli/references/setup.md',
-    ];
-
-    for (const relativePath of firstPartyFiles) {
-      const content = fs.readFileSync(path.join(repoRoot, relativePath), 'utf8');
-      expect(content, relativePath).toContain('bananaslides.online');
-      const contentWithoutProviderLinks = content.replace(/api\.inferera\.com/gi, '');
-      expect(contentWithoutProviderLinks.toLowerCase(), relativePath).not.toContain('inferera.com');
-    }
-
-    const providerExamples = [
-      '.env.example',
-      'docs/configuration.mdx',
-      'docs/zh/configuration.mdx',
-    ];
-
-    for (const relativePath of providerExamples) {
-      const content = fs.readFileSync(path.join(repoRoot, relativePath), 'utf8');
-      expect(content, relativePath).toContain('https://api.inferera.com');
-    }
-  });
 });
